:wrench: refactor (evolve): Merge evolve function and class into one file.'
