+++ pennylane/ops/functions/__init__.py
@@ -31,7 +31,6 @@ This module contains functions that act on operators and tapes.
 """
 from .eigvals import eigvals
 from .equal import equal
-from .evolve import evolve
 from .generator import generator
 from .is_commuting import is_commuting
 from .is_hermitian import is_hermitian
